Support shortcut keys on the SettingsTextBox Form by adding UIActionForm and moving MdiContainerForm.ProcessCmdKey() to it.
